+++ helper_functions/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
-name = "misc"
+name = "helper_functions"
 version = "0.1.0"
-authors = ["Vytautas Krivickas <vytautas1996@gmail.com>"]
+authors = ["BCHF1"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+ring = "0.14.6"
 typenum = '1.11.2'
-types = { path = '../../types' }
+types = { path = '../types' }
