+++ helper_functions/src/crypto.rs
