+++ helper_functions/src/error.rs
