+++ helper_functions/src/predicates.rs
@@ -1,0 +1,167 @@
+pub fn is_slashable_validator(validator: &Validator, epoch: Epoch) -> bool {
+    !validator.slashed && validator.activation_epoch <= epoch && epoch < validator.withdrawable_epoch
+}
+
+pub fn is_active_validator(validator: &Validator, epoch: Epoch) -> bool {
+    validator.activation_epoch <= epoch && epoch < validator.exit_epoch
+}
+
+pub fn is_slashable_attestation_data(data_1: &AttestationData, data_2: &AttestationData) -> bool {
+    // Double vote
+    (data_1 != data_2 && data_1.target.epoch == data_2.target.epoch) ||
+    // Surround vote
+    (data_1.source.epoch < data_2.source.epoch && data_2.target.epoch < data_1.target.epoch)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn is_slashable_validator() {
+        let v = Validator {
+            slashed: false,
+            activation_epoch: 0,
+            withdrawable_epoch: 1,
+            ..Validator::default()
+        };
+        assert_eq!(is_slashable_validator(&v, 0), true);
+    }
+
+    #[test]
+    fn is_slashable_validator_already_slashed() {
+        let v = Validator {
+            slashed: true,
+            activation_epoch: 0,
+            withdrawable_epoch: 1,
+            ..Validator::default()
+        };
+        assert_eq!(is_slashable_validator(&v, 0), false);
+    }
+
+    #[test]
+    fn is_slashable_validator_activation_epoch_greater_than_epoch() {
+        let v = Validator {
+            slashed: false,
+            activation_epoch: 1,
+            withdrawable_epoch: 2,
+            ..Validator::default()
+        };
+        assert_eq!(is_slashable_validator(&v, 0), false);
+    }
+
+    #[test]
+    fn is_slashable_validator_withdrawable_epoch_equals_epoch() {
+        let v = Validator {
+            slashed: false,
+            activation_epoch: 0,
+            withdrawable_epoch: 1,
+            ..Validator::default()
+        };
+        assert_eq!(is_slashable_validator(&v, 1), false);
+    }
+
+    #[test]
+    fn is_active_validator() {
+        let v = Validator {
+            activation_epoch: 0,
+            exit_epoch: 1,
+            ..Validator::default()
+        };
+        assert_eq!(is_active_validator(&v, 0), true);
+    }
+
+    #[test]
+    fn is_active_validator_activation_epoch_greater_than_epoch() {
+        let v = Validator {
+            activation_epoch: 1,
+            exit_epoch: 2,
+            ..Validator::default()
+        };
+        assert_eq!(is_active_validator(&v, 0), false);
+    }
+
+    #[test]
+    fn is_active_validator_exit_epoch_equals_epoch() {
+        let v = Validator {
+            activation_epoch: 0,
+            exit_epoch: 1,
+            ..Validator::default()
+        };
+        assert_eq!(is_active_validator(&v, 1), false);
+    }
+
+    #[test]
+    fn test_is_slashable_attestation_data_double_vote_false() {
+        let attestation_data_1 = AttestationData {
+            target: Checkpoint {
+                epoch: 1,
+                root: H256::from([0; 32]),
+            },
+            ..AttestationData::default()
+        };
+        let attestation_data_2 = AttestationData {
+            target: Checkpoint {
+                epoch: 1,
+                root: H256::from([0; 32]),
+            },
+            ..AttestationData::default()
+        };
+        assert_eq!(
+            is_slashable_attestation_data(&attestation_data_1, &attestation_data_2),
+            false
+        );
+    }
+
+    #[test]
+    fn test_is_slashable_attestation_data_double_vote_true() {
+        let attestation_data_1 = AttestationData {
+            target: Checkpoint {
+                epoch: 1,
+                root: H256::from([0; 32]),
+            },
+            ..AttestationData::default()
+        };
+        let attestation_data_2 = AttestationData {
+            target: Checkpoint {
+                epoch: 1,
+                root: H256::from([1; 32]),
+            },
+            ..AttestationData::default()
+        };
+        assert_eq!(
+            is_slashable_attestation_data(&attestation_data_1, &attestation_data_2),
+            true
+        );
+    }
+
+    #[test]
+    fn test_is_slashable_attestation_data_surround_vote_true() {
+        let attestation_data_1 = AttestationData {
+            source: Checkpoint {
+                epoch: 0,
+                root: H256::from([0; 32]),
+            },
+            target: Checkpoint {
+                epoch: 3,
+                root: H256::from([0; 32]),
+            },
+            ..AttestationData::default()
+        };
+        let attestation_data_2 = AttestationData {
+            source: Checkpoint {
+                epoch: 1,
+                root: H256::from([1; 32]),
+            },
+            target: Checkpoint {
+                epoch: 2,
+                root: H256::from([0; 32]),
+            },
+            ..AttestationData::default()
+        };
+        assert_eq!(
+            is_slashable_attestation_data(&attestation_data_1, &attestation_data_2),
+            true
+        );
+    }
+}
