+++ types/src/types.rs
@@ -40,15 +40,6 @@ pub struct AttestationData {
     pub crosslink: Crosslink,
 }
 
-impl AttestationData {
-    pub fn is_slashable_attestation_data(&self, data: &Self) -> bool {
-        // Double vote
-        (self != data && self.target.epoch == data.target.epoch) ||
-        // Surround vote
-        (self.source.epoch < data.source.epoch && data.target.epoch < self.target.epoch)
-    }
-}
-
 #[derive(
     Clone, PartialEq, Eq, Debug, Deserialize, Serialize, Encode, Decode, TreeHash, SignedRoot,
 )]
@@ -244,169 +235,6 @@ pub struct Validator {
     pub activation_epoch: Epoch,
     pub exit_epoch: Epoch,
     pub withdrawable_epoch: Epoch,
-}
-
-impl Validator {
-    pub fn is_slashable_validator(&self, epoch: Epoch) -> bool {
-        !self.slashed && self.activation_epoch <= epoch && epoch < self.withdrawable_epoch
-    }
-
-    pub fn is_active_validator(&self, epoch: Epoch) -> bool {
-        self.activation_epoch <= epoch && epoch < self.exit_epoch
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn is_slashable_validator() {
-        let v = Validator {
-            slashed: false,
-            activation_epoch: 0,
-            withdrawable_epoch: 1,
-            ..Validator::default()
-        };
-        assert_eq!(v.is_slashable_validator(0), true);
-    }
-
-    #[test]
-    fn is_slashable_validator_already_slashed() {
-        let v = Validator {
-            slashed: true,
-            activation_epoch: 0,
-            withdrawable_epoch: 1,
-            ..Validator::default()
-        };
-        assert_eq!(v.is_slashable_validator(0), false);
-    }
-
-    #[test]
-    fn is_slashable_validator_activation_epoch_greater_than_epoch() {
-        let v = Validator {
-            slashed: false,
-            activation_epoch: 1,
-            withdrawable_epoch: 2,
-            ..Validator::default()
-        };
-        assert_eq!(v.is_slashable_validator(0), false);
-    }
-
-    #[test]
-    fn is_slashable_validator_withdrawable_epoch_equals_epoch() {
-        let v = Validator {
-            slashed: false,
-            activation_epoch: 0,
-            withdrawable_epoch: 1,
-            ..Validator::default()
-        };
-        assert_eq!(v.is_slashable_validator(1), false);
-    }
-
-    #[test]
-    fn is_active_validator() {
-        let v = Validator {
-            activation_epoch: 0,
-            exit_epoch: 1,
-            ..Validator::default()
-        };
-        assert_eq!(v.is_active_validator(0), true);
-    }
-
-    #[test]
-    fn is_active_validator_activation_epoch_greater_than_epoch() {
-        let v = Validator {
-            activation_epoch: 1,
-            exit_epoch: 2,
-            ..Validator::default()
-        };
-        assert_eq!(v.is_active_validator(0), false);
-    }
-
-    #[test]
-    fn is_active_validator_exit_epoch_equals_epoch() {
-        let v = Validator {
-            activation_epoch: 0,
-            exit_epoch: 1,
-            ..Validator::default()
-        };
-        assert_eq!(v.is_active_validator(1), false);
-    }
-
-    #[test]
-    fn test_is_slashable_attestation_data_double_vote_false() {
-        let attestation_data_1 = AttestationData {
-            target: Checkpoint {
-                epoch: 1,
-                root: H256::from([0; 32]),
-            },
-            ..AttestationData::default()
-        };
-        let attestation_data_2 = AttestationData {
-            target: Checkpoint {
-                epoch: 1,
-                root: H256::from([0; 32]),
-            },
-            ..AttestationData::default()
-        };
-        assert_eq!(
-            attestation_data_1.is_slashable_attestation_data(&attestation_data_2),
-            false
-        );
-    }
-
-    #[test]
-    fn test_is_slashable_attestation_data_double_vote_true() {
-        let attestation_data_1 = AttestationData {
-            target: Checkpoint {
-                epoch: 1,
-                root: H256::from([0; 32]),
-            },
-            ..AttestationData::default()
-        };
-        let attestation_data_2 = AttestationData {
-            target: Checkpoint {
-                epoch: 1,
-                root: H256::from([1; 32]),
-            },
-            ..AttestationData::default()
-        };
-        assert_eq!(
-            attestation_data_1.is_slashable_attestation_data(&attestation_data_2),
-            true
-        );
-    }
-
-    #[test]
-    fn test_is_slashable_attestation_data_surround_vote_true() {
-        let attestation_data_1 = AttestationData {
-            source: Checkpoint {
-                epoch: 0,
-                root: H256::from([0; 32]),
-            },
-            target: Checkpoint {
-                epoch: 3,
-                root: H256::from([0; 32]),
-            },
-            ..AttestationData::default()
-        };
-        let attestation_data_2 = AttestationData {
-            source: Checkpoint {
-                epoch: 1,
-                root: H256::from([1; 32]),
-            },
-            target: Checkpoint {
-                epoch: 2,
-                root: H256::from([0; 32]),
-            },
-            ..AttestationData::default()
-        };
-        assert_eq!(
-            attestation_data_1.is_slashable_attestation_data(&attestation_data_2),
-            true
-        );
-    }
 }
 
 #[derive(
