--- utils/misc/src/lib.rs
@@ -1,20 +1,0 @@
-use typenum::marker_traits::Unsigned;
-use types::config::Config;
-use types::primitives::{Epoch, Slot};
-
-pub fn epoch_of_slot<C: Config>(slot: Slot) -> Epoch {
-    slot / C::SlotsPerEpoch::to_u64()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use types::config::MainnetConfig;
-
-    #[test]
-    fn test_epoch_of_slot() {
-        let expected_epoch = 2 as u64;
-        let calculated_epoch = epoch_of_slot::<MainnetConfig>(17 as u64);
-        assert_eq!(calculated_epoch, expected_epoch);
-    }
-}
